Refactor out assignment GR push logic
