+++ app/models/assignment.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 class Assignment < ActiveRecord::Base
-  RELATIONSHIP = 'ministry:relationship'
-
   APPROVED_ROLES = %w(leader inherited_leader admin inherited_admin member).freeze
   LOCAL_APPROVED_ROLES = %w(leader admin member).freeze
   LOCAL_LEADER_ROLES = %w(leader admin).freeze
@@ -76,42 +74,7 @@ class Assignment < ActiveRecord::Base
   end
 
   def create_gr_relationship
-    # Always create the associated global registry entity for assignments with
-    # the root global registry key to prevent a system from gaining
-    # higher-than-expected access to the measurements api data.
-    response = root_gr_client.put(
-      person.gr_id, gr_relationship_entity,
-      params: { full_response: true, fields: RELATIONSHIP }
-    )
-    self.gr_id = relationship_entity_id(response)
-  end
-
-  private
-
-  def root_gr_client
-    GlobalRegistry::Entity.new
-  end
-
-  def gr_relationship_entity
-    {
-      entity: {
-        person: {
-          RELATIONSHIP => {
-            ministry: ministry.gr_id,
-            client_integration_id: "_#{person.gr_id}_#{ministry.gr_id}",
-            team_role: role
-          },
-          'client_integration_id' => ''
-        }
-      }
-    }
-  end
-
-  def relationship_entity_id(response)
-    assignment = response['entity']['person'][RELATIONSHIP].find do |relationship|
-      relationship['ministry'] == ministry.gr_id
-    end
-    assignment['relationship_entity_id']
+    GrSync::AssignmentPush.new(self).push_to_gr
   end
 
   class << self
